crypto-pgpainless: perform hex conversion inside KeyId#toString
